requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
rapidfuzz>=3.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
pyxlsb>=1.0.10
xlrd>=2.0.1
odfpy>=1.4.1
pymongo[srv]>=4.6.0
